Modificação nos testes

+++ src/main/java/br/com/andsantos/northwind/service/impl/CategoriaServiceImpl.java
@@ -36,9 +36,8 @@ public class CategoriaServiceImpl implements CategoriaService {
 		if (repository.existsByNomeCategoria(dto.getNomeCategoria())) {
 			throw new ObjectAlreadyExistsException("Categoria já cadastrada.");
 		}
-		Categoria category = mapper.toEntity(dto);
-        category = repository.save(category);
-        return mapper.toDto(category);
+		Categoria obj = mapper.toEntity(dto);
+        return mapper.toDto(repository.save(obj));
 	}
 
 	@Override
@@ -59,7 +58,11 @@ public class CategoriaServiceImpl implements CategoriaService {
 	@Override
 	public void excluir(Long id) {
         log.debug("Excluindo categoria com id {}", id);
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new NotFoundException("Categoria não encontrada.");
+        }
 	}
 
 	@Override
