+++ src/main/java/br/com/andsantos/northwind/service/impl/ClienteServiceImpl.java
@@ -36,9 +36,9 @@ public class ClienteServiceImpl implements ClienteService {
 		if (repository.existsByNomeEmpresa(dto.getNomeEmpresa())) {
 			throw new ObjectAlreadyExistsException("Cliente já cadastrada.");
 		}
-		Cliente category = mapper.toEntity(dto);
-		category = repository.save(category);
-		return mapper.toDto(category);
+
+		Cliente obj = mapper.toEntity(dto);
+		return mapper.toDto(repository.save(obj));
 	}
 
 	@Override
@@ -56,7 +56,11 @@ public class ClienteServiceImpl implements ClienteService {
 	@Override
 	public void excluir(Long id) {
 		log.debug("Excluindo Cliente com id {}", id);
-		repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new NotFoundException("Cliente não encontrado.");
+        }
 	}
 
 	@Override
