+++ src/main/java/br/com/andsantos/northwind/service/impl/FornecedorServiceImpl.java
@@ -36,9 +36,9 @@ public class FornecedorServiceImpl implements FornecedorService {
 		if (repository.existsByNomeFornecedor(dto.getNomeFornecedor())) {
 			throw new ObjectAlreadyExistsException("Fornecedor já cadastrada.");
 		}
-		Fornecedor category = mapper.toEntity(dto);
-        category = repository.save(category);
-        return mapper.toDto(category);
+
+		Fornecedor obj = mapper.toEntity(dto);
+        return mapper.toDto(repository.save(obj));
 	}
 
 	@Override
@@ -59,7 +59,11 @@ public class FornecedorServiceImpl implements FornecedorService {
 	@Override
 	public void excluir(Long id) {
         log.debug("Excluindo Fornecedor com id {}", id);
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new NotFoundException("Fornecedor não encontrado.");
+        }
 	}
 
 	@Override
